+++ source/layout/svgreplacedbox.h
@@ -33,7 +33,7 @@ public:
     float computeReplacedWidth() const final;
     float computeReplacedHeight() const final;
 
-    void computeIntrinsicRatioInformation(float& intrinsicWidth, float& intrinsicHeight, double& intrinsicRatio) const final;
+    void computeIntrinsicRatioInformation(float& intrinsicWidth, float& intrinsicHeight, float& intrinsicRatio) const final;
     void paintReplaced(const PaintInfo& info, const Point& offset) final;
     void layout(FragmentBuilder* fragmentainer) final;
     void build() final;
